refactor(audit): extract release-train primitives into a leaf core module
`stripV`, `compareSemver`, `ageHours`, `decideExit` and the `EdgeResult` /
`ReleaseInfo` / `PublishedRelease` vocabulary move verbatim into
`_release-train-core.ts`, re-exported from `check-release-staleness.ts` so the
module path callers use is unchanged. The leaf module exists so the Phase 2
dependency readers can share these without forming an import cycle with the
Phase 1 file.

`ageHours` additionally now fails closed on a timestamp carrying no explicit
zone: `new Date("2026-07-26T12:00:00")` parses as LOCAL time, so the age would
differ by up to 14h between a laptop and a UTC runner — enough to move an edge
across the 6h grace boundary. Both real sources are Z-suffixed, so no production
edge changes.

39 tests before the move, 39 after (proving it verbatim), 41 with the two
timezone guards.
